Added: -ClientPlayground class, where user code for setting up rendering will now be located.
Changed:
-Updated allocator to include debug memory tracking and memory leak dumping.
-Draw Batch now no longer expects an MVP buffer, but instead a binding layout in which the user can provide any additional bindings to the draw batch pipeline.

Fixed:
-Draw batch shader would not draw 256 objects, since the instance index was read in as a signed char instead of unsigned.

+++ VKR/VKRClient/TestAssets/Shaders/GLSL/vs_obj_def_batch.vert
@@ -62,8 +62,8 @@ layout(location = 0) out VS_OUT
 
 void main() 
 {
-    uint vertexIndex = gl_VertexIndex & 0xFFFFFF; // Mask out final 8 bits for vertex index.
-    uint instanceIndex = gl_VertexIndex >> 24; // Shift 24 bits right for instance index.
+    uint vertexIndex = uint(gl_VertexIndex) & 0xFFFFFF; // Mask out final 8 bits for vertex index.
+    uint instanceIndex = uint(gl_VertexIndex) >> 24; // Shift 24 bits right for instance index.
 
     VS_IN vsInput = vertexBuffers[nonuniformEXT(bindings.vertexBufferIndex)].vertices[nonuniformEXT(vertexIndex)];
     VS_INSTANCE vsInstance = instanceBuffers[nonuniformEXT(bindings.instanceBufferIndex)].instances[nonuniformEXT(instanceIndex)];
